Opprydding iht. kommentarer + saksbehandler som oppgavekilde
