+++ libs/etterlatte-oppgave-model/src/main/kotlin/oppgave/OppgaveIntern.kt
@@ -103,6 +103,7 @@ enum class OppgaveKilde {
     EKSTERN,
     TILBAKEKREVING,
     GJENOPPRETTING,
+    SAKSBEHANDLER,
 }
 
 enum class OppgaveType {
